requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.2
djangorestframework==3.14.0
docopt==0.6.2
idna==3.6
Js2Py==0.74
packaging==23.2
paho-mqtt==1.6.1
pipwin==0.5.2
psycopg==3.1.18
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
